Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pyodbc
python-dotenv
python-docx
pandas